Add a Python version, make final tweaks to the bash one

+++ build-openmw.sh
@@ -1,13 +1,10 @@
-#!/bin/bash -e
-
+#!/bin/bash -ex
 # https://wiki.openmw.org/index.php?title=Development_Environment_Setup
-
 build_cpus=$(cat /proc/cpuinfo | grep proc | wc -l)
 install_prefix=/opt/morrowind
 source_dir=${HOME}/src
-
 bullet3_version=2.86.1
-ffmpeg_version=2.8.11
+ffmpeg_version=2.8.12
 mygui_version=3.2.2
 unshield_version=1.4.2
 
@@ -23,7 +20,7 @@ function help-text
 {
     cat <<EOF
 
-Usage: build-openmw [-h] [-J] [-M] [-N] [-s SHA] [-t TAG] [--with-mygui] [--with-unshield]
+Usage: build-openmw [-h] [-J] [-M] [-N] [-I DIR] [-s SHA] [-t TAG] [--force]
 
 Build OpenMW!
 
@@ -32,14 +29,13 @@ into a tarball along with its dependencies.
 
 Optional Arguments:
   --force               Force building, even if the requested revision is already built
-  -h, --help            Show this help message and exit
+  -I, --install-dir     Compiled files will be installed to.  Default: ${install_prefix}
   -J, --just-openmw     Only package OpenMW
+  -h, --help            Show this help message and exit
   -M, --mp              Build TES3MP (Multiplayer fork of OpenMW - EXPERIMENTAL)
-  -N, --no-tar          Don't create any tarballs
+  -N, --no-pkg          Don't create any tarballs
   -s SHA, --sha SHA     Build the specified git revision (sha1)
   -t TAG, --tag TAG     Build the specified git tag
-  --with-mygui          Build MyGUI and link against it
-  --with-unshield       Build Unshield and link against it
 
 EOF
     exit 1
@@ -63,9 +59,19 @@ function install-pkgs-void
 {
     echo VOID
     export distro=void
-    # TODO: need MUCH better error handling here... maybe check xbps-install's output
-    sudo xbps-install -y boost-devel cmake freetype-devel gcc git libmygui-devel libopenal-devel libopenjpeg2-devel libtxc_dxtn libunshield-devel libXt-devel make nasm ois-devel python-devel python3-devel qt-devel qt5-devel SDL2-devel zlib-devel \
-        || echo "DIDN'T INSTALL PACKAGES -- DO WE EVEN NEED THEM??"
+    # if ${multiplayer}; then
+    #     # TODO: need MUCH better error handling here... maybe check xbps-install's output
+    #     sudo xbps-install -y boost-devel cmake freetype-devel gcc git libmygui-devel libopenal-devel libopenjpeg2-devel libtxc_dxtn libunshield-devel libXt-devel make nasm ois-devel python-devel python3-devel qt-devel SDL2-devel zlib-devel \
+    #         || echo "DIDN'T INSTALL PACKAGES -- DO WE EVEN NEED THEM??"
+    # else
+    clear
+    echo
+    echo ENTER SUDO PASSWORD TO INSTALL PACKAGES...
+# WARNING: pkg-config not found, library detection may fail.
+    echo
+        sudo xbps-install -y boost-devel cmake freetype-devel gcc git libmygui-devel libopenal-devel libopenjpeg2-devel libtxc_dxtn libunshield-devel libXt-devel libopenal-devel make nasm ois-devel python-devel python3-devel qt-devel SDL2-devel zlib-devel \
+            || echo "DIDN'T INSTALL PACKAGES -- DO WE EVEN NEED THEM??"
+    # fi
 }
 
 function install-pkgs
@@ -89,6 +95,10 @@ function fix-libs
     # -- Installing: /opt/morrowind/openmw-cb32f1d60/share/games/openmw/resources/defaultfilters
     # -- Installing: /opt/morrowind/openmw-cb32f1d60/share/games/openmw/data
     # chmod: cannot access '/opt/morrowind/openmw-cb32f1d60/bin/*': No such file or directory
+    clear
+    echo
+    echo ENTER SUDO PASSWORD TO FIX EXECUTABLES...
+    echo
     sudo find ${install_prefix} -type d -exec chmod 0755 {} \; 2>/dev/null
     sudo find ${install_prefix} -type f -exec chmod 0644 {} \; 2>/dev/null
 }
@@ -109,7 +119,7 @@ function build-ffmpeg2
         git checkout n${ffmpeg_version}
         ./configure --prefix=${install_prefix}/ffmpeg-${ffmpeg_version}
         make -j${build_cpus}
-        sudo make install
+        make install
         fix-libs
     else
         echo "FOUND!"
@@ -133,7 +143,7 @@ function build-osg-openmw
         git pull  # we always want the bleeding edge for this one.
         cmake -D CMAKE_INSTALL_PREFIX=${install_prefix}/osg-openmw ..
         make -j${build_cpus}
-        sudo make install
+        make install
         fix-libs
     else
         echo "FOUND!"
@@ -143,7 +153,7 @@ function build-osg-openmw
 function build-bullet
 {
     printf "Checking for bullet... "
-    if ! [ -f ${install_prefix}/bullet3-${bullet3_version}/lib/libLinearMath.a ]; then
+    if ! [ -f ${install_prefix}/bullet3-${bullet3_version}/lib/libLinearMath.so ]; then
         echo "NOT found!  building..."
         if ! [ -d ${source_dir}/bullet3 ]; then
             cd ${source_dir}
@@ -159,9 +169,11 @@ function build-bullet
               -D BUILD_OPENGL3_DEMOS=false \
               -D BUILD_BULLET2_DEMOS=false \
               -D BUILD_UNIT_TESTS=false \
+              -D INSTALL_LIBS=on \
+              -D BUILD_SHARED_LIBS=on \
               ..
         make -j${build_cpus}
-        sudo make install
+        make install
         fix-libs
     else
         echo "FOUND!"
@@ -186,12 +198,12 @@ function build-unshield
         # cp ../../unshield/build/libunshield.pc /usr/local/lib/pkgconfig/
         cmake -D CMAKE_INSTALL_PREFIX=${install_prefix}/unshield-${unshield_version} ..
         make -j${build_cpus}
-        sudo make install
+        make install
         fix-libs
     else
         echo "FOUND!"
     fi
-    sudo chmod 0755 ${install_prefix}/unshield-${unshield_version}/bin/*
+    chmod 0755 ${install_prefix}/unshield-${unshield_version}/bin/*
 }
 
 function build-mygui
@@ -209,6 +221,8 @@ function build-mygui
         git checkout -- ..
         git clean -df
         git checkout MyGUI${mygui_version}
+        # git checkout master
+        # git reset --hard origin/master
         cmake -D CMAKE_INSTALL_PREFIX=${install_prefix}/mygui-${mygui_version} \
               -D MYGUI_RENDERSYSTEM=1 \
               -D MYGUI_BUILD_DEMOS=OFF \
@@ -216,7 +230,7 @@ function build-mygui
               -D MYGUI_BUILD_PLUGINS=OFF \
               ..
         make -j${build_cpus}
-        sudo make install
+        make install
         fix-libs
     else
         echo "FOUND!"
@@ -300,9 +314,11 @@ function build-terra
 
 function export-openmw-sha
 {
-    cd ${source_dir}/openmw
-    export openmw_sha=$(git rev-parse HEAD)
-    export openmw_sha_short=$(git rev-parse --short HEAD)
+    if [ -d ${source_dir}/openmw ]; then
+        cd ${source_dir}/openmw
+        export openmw_sha=$(git rev-parse HEAD)
+        export openmw_sha_short=$(git rev-parse --short HEAD)
+    fi
 }
 
 function export-app-sha
@@ -347,11 +363,12 @@ function build-tes3mp
             git clone ${_clone_url} ${_app}
         fi
         export-app-sha ${_app}
-        _cmake_params="-Wno-dev -D CMAKE_BUILD_TYPE=Release -D BUILD_OPENCS=OFF -D CMAKE_CXX_STANDARD=14 -D CallFF_INCLUDES=${install_prefix}/callff/include -D CallFF_LIBRARY=${install_prefix}/callff/libcallff.a -D RakNet_INCLUDES=${install_prefix}/raknet/include -D RakNet_LIBRARY_DEBUG=${install_prefix}/raknet/libRakNetLibStatic.a -D RakNet_LIBRARY_RELEASE=${install_prefix}/raknet/libRakNetLibStatic.a -D Terra_INCLUDES=${install_prefix}/terra/include -D Terra_LIBRARY_RELEASE=${install_prefix}/terra/lib/libterra.a"
+        # _cmake_params="-Wno-dev -D CMAKE_BUILD_TYPE=Release -D BUILD_OPENCS=OFF -D CMAKE_CXX_STANDARD=14 -D CallFF_INCLUDES=${install_prefix}/callff/include -D CallFF_LIBRARY=${install_prefix}/callff/libcallff.a -D RakNet_INCLUDES=${install_prefix}/raknet/include -D RakNet_LIBRARY_DEBUG=${install_prefix}/raknet/libRakNetLibStatic.a -D RakNet_LIBRARY_RELEASE=${install_prefix}/raknet/libRakNetLibStatic.a -D Terra_INCLUDES=${install_prefix}/terra/include -D Terra_LIBRARY_RELEASE=${install_prefix}/terra/lib/libterra.a"
+        _cmake_params="-Wno-dev -D CMAKE_BUILD_TYPE=Release -D BUILD_OPENCS=OFF -D CMAKE_CXX_STANDARD=14 -D CallFF_INCLUDES=${install_prefix}/callff/include -D CallFF_LIBRARY=${install_prefix}/callff/libcallff.a -D RakNet_INCLUDES=${install_prefix}/raknet/include -D RakNet_LIBRARY_DEBUG=${install_prefix}/raknet/libRakNetLibStatic.a -D RakNet_LIBRARY_RELEASE=${install_prefix}/raknet/libRakNetLibStatic.a -D FORCE_LUA=1 -D LUA_INCLUDE_DIR=/usr/include/lua5.1"
         if ! [ -d ${install_prefix}/${_app}-${app_sha_short} ] || ${force}; then
             if ${force}; then
                 echo "${_app}-${app_sha_short} exists but requested to force!  FORCING!!!!"
-                sudo /bin/rm -rf ${_app}-${app_sha_short}
+                /bin/rm -rf ${_app}-${app_sha_short}
             else
                 echo "${_app} NOT found!  building..."
             fi
@@ -371,20 +388,26 @@ function build-tes3mp
             fi
             if [ $distro = void ] && ! [ -h /usr/lib/libtinfo.so.6.0 ] || ! [ -h /usr/lib/libtinfo.so ]; then
                 echo MAKING libtinfo COMPAT SYMLINK ........................
+                clear
+                echo
+                echo ENTER SUDO PASSWORD TO MAKE libtinfo SYMLINK...
+                echo
                 sudo ln -fs /usr/lib/libncurses.so.6.0 /usr/lib/libtinfo.so.6.0
                 sudo ln -fs /usr/lib/libtinfo.so.6.0 /usr/lib/libtinfo.so
             fi
             mkdir -p ${install_prefix}/${_app}-${app_sha_short}
-            export LDFLAGS="-lz -lbz2 -lncurses"
+            export LDFLAGS="-lz -llzma -lbz2 -lncurses"
             cmake ${_cmake_params} ..
             make -j${build_cpus}
             /bin/cp -fr apps components docs extern files resources bsatool esmtool gamecontrollerdb.txt openmw-cs.cfg openmw-essimporter openmw-iniimporter openmw-launcher openmw-wizard openmw.appdata.xml openmw.cfg openmw.desktop settings-default.cfg tes3mp tes3mp-browser tes3mp-client-default.cfg tes3mp-server tes3mp-server-default.cfg ${install_prefix}/${_app}-${app_sha_short}/
             fix-libs
-            sudo sh -c "echo ${app_sha} > ${install_prefix}/${_app}-${app_sha_short}/REVISION"
-            if [ ${_app} = tes3mp ] && ! [ -d ${install_prefix}/${_app}-${app_sha_short}/PluginExamples ]; then
-                echo ADDING PLUGIN EXAMPLES....................
-                cd ${install_prefix}/${_app}-${app_sha_short}
-                git clone https://github.com/TES3MP/PluginExamples.git
+            sh -c "echo ${app_sha} > ${install_prefix}/${_app}-${app_sha_short}/REVISION"
+            if [ ${_app} = tes3mp ]; then
+                if ! [ -d ${install_prefix}/${_app}-${app_sha_short}/PluginExamples ]; then
+                    echo ADDING PLUGIN EXAMPLES....................
+                    cd ${install_prefix}/${_app}-${app_sha_short}
+                    git clone https://github.com/TES3MP/PluginExamples.git
+                fi
                 sed -i "s|home = .*|home = ./PluginExamples|g" tes3mp-server-default.cfg
             fi
         else
@@ -398,7 +421,7 @@ function build-tes3mp
             /bin/rm -f ${install_prefix}/openmw
         fi
         _user=$(whoami)
-        sudo chown -R ${_user}: ${install_prefix}
+        chown -R ${_user}: ${install_prefix}
         cd ${install_prefix}
         ln -s openmw-${openmw_sha_short} openmw
     else
@@ -406,7 +429,7 @@ function build-tes3mp
             /bin/rm -f ${install_prefix}/${_app}
         fi
         _user=$(whoami)
-        sudo chown -R ${_user}: ${install_prefix}
+        chown -R ${_user}: ${install_prefix}
         cd ${install_prefix}
         ln -s ${_app}-${app_sha_short} ${_app}
     fi
@@ -415,16 +438,18 @@ function build-tes3mp
 function build-openmw
 {
     printf "Checking for openmw... "
-    cd ${source_dir}/openmw
-    git checkout master
-    git checkout -- .
-    git clean -df
-    git pull --rebase
-    if ! [ -z "${1}" ]; then
-        git checkout "${1}"
+    if [ -d ${source_dir}/openmw ]; then
+        cd ${source_dir}/openmw
+        git checkout master
+        git checkout -- .
+        git clean -df
+        git pull --rebase
+        if ! [ -z "${1}" ]; then
+            git checkout "${1}"
+        fi
+        # Re-export the sha in case upstream changes have been pulled down
+        export-openmw-sha
     fi
-    # Re-export the sha in case upstream changes have been pulled down
-    export-openmw-sha
     if ! [ -f ${install_prefix}/openmw-${openmw_sha_short}/bin/openmw ] || ${force}; then
         if ! [ -d ${source_dir}/openmw ]; then
             cd ${source_dir}
@@ -433,7 +458,7 @@ function build-openmw
         if ! [ -d ${install_prefix}/openmw-${openmw_sha_short} ] || ${force}; then
             if ${force}; then
                 echo 'openmw-${openmw_sha_short} exists but requested to force!'
-                sudo /bin/rm -rf ${install_prefix}/openmw-${openmw_sha_short}
+                /bin/rm -rf ${install_prefix}/openmw-${openmw_sha_short}
             else
                 echo "NOT found!  building..."
             fi
@@ -443,7 +468,7 @@ function build-openmw
             # if [ ${code} -eq 0 ] && [ "${current_install}" != "openmw-${openmw_sha_short}" ]; then
             #     # We are building a new sha, so remove the old one
             #     current_install=$(ls -1d ${install_prefix}/openmw-* | awk -F/ '{ print $4 }' 2>/dev/null)
-            #     sudo rm -rf /opt/morrowind/${current_install}
+            #     rm -rf /opt/morrowind/${current_install}
             # fi
             [ -d ${source_dir}/openmw/build ] && rm -rf ${source_dir}/openmw/build
             mkdir -p ${source_dir}/openmw/build
@@ -457,12 +482,12 @@ function build-openmw
             else
                 export CMAKE_PREFIX_PATH=${install_prefix}/ffmpeg-${ffmpeg_version}:${install_prefix}/osg-openmw:${install_prefix}/bullet3-${bullet3_version}
             fi
-            export LDFLAGS="-lz -lbz2"
+            export LDFLAGS="-llzma -lz -lbz2"
             cmake -D CMAKE_BUILD_TYPE=Release -D CMAKE_INSTALL_PREFIX=${install_prefix}/openmw-${openmw_sha_short} ..
             make -j${build_cpus}
-            sudo make install
+            make install
             fix-libs
-            sudo sh -c "echo ${openmw_sha} > ${install_prefix}/openmw-${openmw_sha_short}/REVISION"
+            sh -c "echo ${openmw_sha} > ${install_prefix}/openmw-${openmw_sha_short}/REVISION"
         else
             echo "FOUND!"
         fi
@@ -473,24 +498,28 @@ function build-openmw
         /bin/rm -f ${install_prefix}/openmw
     fi
     _user=$(whoami)
-    sudo chown -R ${_user}: ${install_prefix}
+    chown -R ${_user}: ${install_prefix}
     cd ${install_prefix}
     ln -s openmw-${openmw_sha_short} openmw
 }
 
 function make-bins-executable
 {
-    sudo chmod 0755 ${install_prefix}/ffmpeg-${ffmpeg_version}/bin/*
-    sudo chmod 0755 ${install_prefix}/osg-openmw/bin/*
-    sudo find ${install_prefix}/openmw-${openmw_sha_short}/bin -exec chmod 0755 {} \;
+    chmod 0755 ${install_prefix}/ffmpeg-${ffmpeg_version}/bin/*
+    chmod 0755 ${install_prefix}/osg-openmw/bin/*
+    find ${install_prefix}/openmw-${openmw_sha_short}/bin -exec chmod 0755 {} \;
 }
 
 function tar-it-up
 {
     version="${1}"
-    filename=morrowind-${distro}-${version}.tar.bzip2
+    filename=morrowind-${distro}-${version}.tar.bz2
     printf "Creating ${filename} ... "
     cd /opt
+    clear
+    echo
+    echo ENTER SUDO PASSWORD TO CLEAN AND/OR CREATE A PACKAGE...
+    echo
     [ -f ${filename} ] && sudo rm -rf ${filename}
     sudo tar cjpf ${filename} morrowind
     sudo chown $(whoami): ${filename}
@@ -502,7 +531,7 @@ function tar-the-thing
 {
     filename="${1}"
     [[ -z ${filename} ]] && error-and-die "Need to specify a file name for 'tar-the-thing'!"
-    tarname=${filename}.tar.bzip2
+    tarname=${filename}.tar.bz2
     printf "Creating ${tarname} ... "
     cd ${install_prefix}
     [ -f ${tarname} ] && sudo rm -rf ${tarname}
@@ -517,12 +546,10 @@ function main
     force=false
     just_openmw=false
     export multiplayer=false
-    notar=false
+    nopkg=false
     sha=false
     tag=false
-    export with_mygui=false
-    export with_unshield=false
-    opts=$(getopt -o JMNhs:t: --longoptions force,help,just-openmw,mp,no-tar,sha:,tag:,with-mygui,with-unshield -n build-openmw -- "${@}")
+    opts=$(getopt -o I:JMNhs:t: --longoptions force,help,install-dir:,just-openmw,mp,no-pkg,sha:,tag: -n build-openmw -- "${@}")
 
     eval set -- "$opts"
 
@@ -530,39 +557,37 @@ function main
         case "$1" in
             --force ) force=true; shift;;
             -h | --help ) help-text; shift;;
+            -I | --install-dir ) export install_prefix="$2"; shift;;
             -s | --sha ) sha=true; openmw_build_sha="$2"; shift; shift ;;
             -t | --tag ) tag=true; openmw_build_tag="$2"; shift; shift ;;
-            # -L | --just-lib ) just_lib="${2}"; shift; shift ;;
             -J | --just-openmw ) just_openmw=true; shift ;;
             -M | --mp ) multiplayer=true; shift ;;
-            -N | --no-tar ) notar=true; shift ;;
-            --with-mygui ) with_mygui=true; shift ;;
-            --with-unshield ) with_unshield=true; shift ;;
+            -N | --no-pkg ) nopkg=true; shift ;;
             -- ) shift; break ;;
             * ) break ;;
         esac
     done
 
+    if ! [ -d ${install_prefix} ]; then
+        sudo mkdir ${install_prefix}
+        sudo chown -R $(whoami) ${install_prefix}
+    fi
+
     mkdir -p ${source_dir}
+    # sudo chown -R $(whoami): ${source_dir}
     install-pkgs
 
     build-ffmpeg2
     build-osg-openmw
     build-bullet
-    if ${with_mygui}; then
-        echo WITH UNSHIELD ----------
-        build-unshield
-    fi
-    if ${with_unshield}; then
-        echo WITH MYGUI ----------
-        build-mygui
-    fi
+    build-unshield
+    build-mygui
 
     if ${multiplayer}; then
         build-callff
         build-raknet
         # TODO: build against lua, not terra
-        build-terra
+        # build-terra
         build-tes3mp
         echo "MULTIPLAYER DONE!!!!"
         exit 0
@@ -583,9 +608,11 @@ function main
         fi
     fi
 
+    # TODO: make standalone package
+
     make-bins-executable
     # Don't create any package
-    if ${notar}; then
+    if ${nopkg}; then
         echo 'No-tar option specified; not tarring!'
     # Only package OpenMW
     elif ${just_openmw}; then
@@ -594,6 +621,7 @@ function main
     # Create the full runtime package
     else
         tar-it-up ${version}
+        # make-standalone-package ${version}
     fi
 }
